🎨 supabase 쿼리 코드 개선
