+++ src/utils/supabase/constant.ts
@@ -1,0 +1,1 @@
+export const RESUME_TABLE = "resume";
